Per-Checker/requirements.txt
psutil==5.9.0
GPUtil==1.4.0
speedtest-cli==2.1.3
colorama==0.4.6